[ar 2.0] Add URI resolver feature
This feature provides native support for dispatching
to different ArResolver implementations based on URI
scheme. This allows users to plug-in support for asset
systems without changing the "primary" resolver
implementation.

Creating a new URI resolver involves implementing a
subclass of ArResolver and then specifying the URI
schemes it handles via plugin metadata. For example:

"Types": {
    "HTTPResolver": {
        "bases": ["ArResolver"],
        "uriSchemes": ["http", "https"]
     }
}

In this example, the HTTPResolver implementation is
registered to handle the "http" and "https" schemes.
Any queries involving http:... or https:... URIs will
automatically be dispatched to the corresponding
function on HTTPResolver.

(Internal change: 2121003)
(Internal change: 2121280)
(Internal change: 2123206)