[ar 2.0] Add ability to create ArResolverContext from strings
Clients can use the new ArResolver::CreateContextFromString[s]
API to create ArResolverContext objects from a generic string.
The format of the string is opaque to Ar itself. Internally,
Ar will forward the context string to the underlying resolver
implementation and allow it to parse and return the appropriate
context object.

ArDefaultResolver supports creating an ArDefaultResolverContext
whose search path is specified by a ":"-delimited list of paths
(";" on Windows).

This provides clients a generic way to create an ArResolverContext.
In particular, this is useful for DCCs or plugins that want to
allow clients to specify an ArResolverContext to bind when
opening a UsdStage.

(Internal change: 2124809)