[ar 2.0] Add API for writing assets
This change introduces the ArWritableAsset interface, which
is similar to the ArAsset interface but used for writing
instead of reading data. Users can implement this interface
and the ArResolver::OpenAssetForWrite function to provide
a way for USD to write data to whatever storage is backing
their asset system.

ArDefaultResolver has been updated to accommodate this new
API by using an ArFilesystemWritableAsset. This is a
filesystem-based implementation that uses a TfSafeOutputFile
to write to a temporary file before renaming to the final
destination location. This is the same scheme that is
currently used to write .usda and .usdc files. Follow-up
changes will update the .usda and .usdc file formats to
use the ArWritableAsset API instead of their current
filesystem-specific mechanisms.

(Internal change: 2130054)