[ar 2.0] (Re)introduce API for computing asset info
Change 2128577 removed ArResolver::UpdateAssetInfo and
made Resolve/ResolveForNewAsset responsible for computing
an asset's ArAssetInfo. Per the Ar 2.0 proposal, this
was intended to simplify the ArResolver interface.
Computing the ArAssetInfo as part of resolution seemed
reasonable, since that information would probably be
available at that time.

However, subsequent performance tests revealed a flaw
in this approach: SdfLayer resolves asset paths and
computes the asset info as part of layer lookup in
SdfLayer::Find/FindOrOpen, even though the asset info
is only needed when actually opening a layer for the
first time. This meant SdfLayer was doing unnecessary
(and potentially) costly work for layers that were
already opened.

One possible approach to fixing this issue would be
to resolve the asset path *without* computing the
asset info when looking up the layer, then re-resolve
the path and compute the asset info when SdfLayer
determined it was opening a layer for the first time.
This would not require changing the ArResolver API.
However, this would penalize all users by requiring
two resolves when opening an SdfLayer, including those
who don't use ArAssetInfo to store asset information.
This was reason enough to reject this solution.

Instead, this change removes the responsibility for
computing the asset info from Resolve and moves it
to a new method, ArResolver::GetAssetInfo. This
simplifies the Resolve API and allows SdfLayer to
compute the asset info only when it knows it's
needed.

(Internal change: 2130328)