[ar 2.0] Remove deprecated ArResolver::ConfigureResolverForAsset
This method was introduced during the initial refactoring of
USD prior to its open source release. Tools like usdview had
been using Pixar-specific APIs to configure the in-house resolver
at startup, and a plugin point was needed to maintain this
behavior while removing the Pixar-specific code to allow these
tools to be released publicly.

Like other vestigial Ar APIs, this method was not well-documented.
It was unclear to clients when it should be called, what its
effect would be, how it would affect previously-resolved paths
or what it would/should do in applications that dealt with
multiple assets. Conceptually, the ArResolver is a global
resource that USD may be using to handle one asset or
multiple assets, potentially at the same time. With that
perspective, having a configuration method for a single asset
doesn't make sense.

ArResolver provides the resolver context mechanism as a way
for clients to encode asset-specific resolution behaviors.
With Ar 2.0, new methods were added to create resolver contexts
from arbitrary strings. This will hopefully be able to replace
whatever work was being done by implementations in the
removed function.

(Internal change: 2175388)